improved modules isolation
