+++ connectome/containers/__init__.py
@@ -1,0 +1,3 @@
+from .base import *
+from .context import *
+from .reversible import ReversibleContainer
