+++ connectome/containers/reversible.py
@@ -1,16 +1,15 @@
+from typing import Tuple
+
+from . import BagContext
+from ..engine import Nodes, BoundEdges
 from .base import EdgesBag
-from .context import BagContext
-from ..engine.base import Nodes, BoundEdges
-from ..interface.factory import normalize_inherit
-from ..layers.chain import connect, ChainContext
-from ..utils import check_for_duplicates, node_to_dict, deprecation_warn, NameSet
+from ..utils import NameSet, node_to_dict, check_for_duplicates, AntiSet
 
 
-class TransformContainer(EdgesBag):  # pragma: no cover
+class ReversibleContainer(EdgesBag):
     def __init__(self, inputs: Nodes, outputs: Nodes, edges: BoundEdges, backward_inputs: Nodes = (),
                  backward_outputs: Nodes = (), *, optional_nodes: NameSet = None,
                  forward_virtual: NameSet, backward_virtual: NameSet, persistent_nodes: NameSet = None):
-        deprecation_warn()
         forward_virtual, valid = normalize_inherit(forward_virtual, node_to_dict(outputs))
         assert valid
         backward_virtual, valid = normalize_inherit(backward_virtual, node_to_dict(backward_outputs))
@@ -24,8 +23,18 @@ class TransformContainer(EdgesBag):  # pragma: no cover
             optional_nodes=optional_nodes,
         )
 
-    def wrap(self, container: 'EdgesBag') -> 'EdgesBag':
-        return connect(container, self)
 
+def normalize_inherit(value, outputs) -> Tuple[NameSet, bool]:
+    if isinstance(value, str):
+        value = [value]
 
-PipelineContext = ChainContext
+    if isinstance(value, bool):
+        valid = value
+        value = AntiSet(set(outputs))
+    elif isinstance(value, AntiSet):
+        valid = True
+    else:
+        value = set(value)
+        valid = all(isinstance(node_name, str) for node_name in value)
+
+    return value, valid
