+++ connectome/interface/factory.py
@@ -1,17 +1,18 @@
 import inspect
 import logging
-from typing import Dict, Any, Iterable, Callable, Type, Tuple, Set
+from typing import Dict, Any, Iterable, Callable, Type
 
+from .base import CallableLayer
+from .decorators import Meta, Optional, RuntimeAnnotation
 from .edges import EdgeFactory, Function
-from ..containers.base import EdgesBag, BagContext
-from ..engine.edges import IdentityEdge, ConstantEdge
-from ..exceptions import GraphError, FieldError
-from ..utils import MultiDict, check_for_duplicates, node_to_dict, AntiSet
+from .factory_utils import add_quals, to_argument
 from .nodes import NodeStorage, Input, InverseInput, Parameter, InverseOutput, Output, NodeTypes, NodeType, Default, \
     Intermediate, FinalNodeType, is_private
-from .base import CallableLayer
-from .factory_utils import add_quals, to_argument
-from .decorators import Meta, Optional, RuntimeAnnotation
+from ..containers import EdgesBag, ReversibleContainer
+from ..containers.reversible import normalize_inherit
+from ..engine import IdentityEdge, ConstantEdge
+from ..exceptions import GraphError, FieldError
+from ..utils import MultiDict
 
 logger = logging.getLogger(__name__)
 
@@ -313,24 +314,12 @@ class GraphFactory:
             'Compiling layer. Inputs: %s, Outputs: %s, BackwardInputs: %s, BackwardOutputs: %s',
             list(self.inputs), list(self.outputs), list(self.backward_inputs), list(self.backward_outputs),
         )
-
-        inputs = list(self.inputs.values())
-        outputs = list(self.outputs.values())
-        backward_inputs = list(self.backward_inputs.values())
-        backward_outputs = list(self.backward_outputs.values())
-        edges = self.edges + list(self._get_constant_edges(arguments))
-
-        forward_virtual, valid = normalize_inherit(self.forward_inherit, node_to_dict(outputs))
-        assert valid
-        backward_virtual, valid = normalize_inherit(self.backward_inherit, node_to_dict(backward_outputs))
-        assert valid
-
-        check_for_duplicates(inputs)
-        return EdgesBag(
-            inputs, outputs, edges,
-            BagContext(backward_inputs, backward_outputs, backward_virtual),
-            virtual_nodes=forward_virtual, persistent_nodes=self.persistent_names,
-            optional_nodes=self.optional_names,
+        return ReversibleContainer(
+            list(self.inputs.values()), list(self.outputs.values()),
+            self.edges + list(self._get_constant_edges(arguments)),
+            list(self.backward_inputs.values()), list(self.backward_outputs.values()),
+            optional_nodes=self.optional_names, forward_virtual=self.forward_inherit,
+            backward_virtual=self.backward_inherit, persistent_nodes=self.persistent_names,
         )
 
 
@@ -406,19 +395,3 @@ def items_to_container(items, inherit, factory_cls: Type[GraphFactory]):
     if factory.special_methods:
         raise TypeError(f"This constructor doesn't accept special methods: {tuple(factory.special_methods)}")
     return factory.build({}), factory.property_names
-
-
-def normalize_inherit(value, outputs) -> Tuple[Set[str], bool]:
-    if isinstance(value, str):
-        value = [value]
-
-    if isinstance(value, bool):
-        valid = value
-        value = AntiSet(set(outputs))
-    elif isinstance(value, AntiSet):
-        valid = True
-    else:
-        value = set(value)
-        valid = all(isinstance(node_name, str) for node_name in value)
-
-    return value, valid
